requirements.txt
echo "openai
tiktoken
faiss-cpu
jieba" > requirements.txt
